Bringing over a needed change we missed in the last update
We no longer are going to attempt to change the extension of a file out from under the user of a dialog. This implementation is problematic for a variety of reasons. If a similar behavior is ever needed (because the exact behavior is wrong), then another implementation strategy should be used.